+++ src/odr/internal/html/xml_file.cpp
@@ -259,7 +259,7 @@ public:
     out.write_header_target("_blank");
     out.write_header_title("odr");
     write_viewport_meta(out, config(), false);
-    write_zoom_style(out, config(), false, {});
+    write_zoom_style(out, config(), WidthFit::none, {});
 
     write_xml_style(state);
     write_xml_dark_style(state);